+++ DalaMock.Sample/Mock/Program.cs
@@ -1,3 +1,9 @@
+using System;
+using System.Collections.Generic;
+
+using DalaMock.Core.Mocks;
+using DalaMock.Sample.Services;
+
 namespace DalaMock.Sample.Mock;
 
 using DalaMock.Core.DI;
@@ -6,7 +12,10 @@ internal static class Program
 {
     private static void Main(string[] args)
     {
-        var mockContainer = new MockContainer();
+        var mockContainer = new MockContainer(serviceReplacements: new Dictionary<Type, Type>()
+        {
+            { typeof(MockPluginLog), typeof(MockPluginLogReplacement) },
+        });
         var mockDalamudUi = mockContainer.GetMockUi();
         var pluginLoader = mockContainer.GetPluginLoader();
         var mockPlugin = pluginLoader.AddPlugin(typeof(DalamudMockPluginTest));